requirements.txt
future==0.16.0
PyYAML==3.12
click==6.7
troposphere==2.3.1
awacs==0.7.2
boto3==1.7.37
halo==0.0.12
kubernetes==6.0.0
oyaml==0.4
jinja2==2.10
netaddr==0.7.19
tabulate==0.8.2
